moved searchbar inside map
